backend/requirements.txt
# Python >= 3.10 required
# Canonical deps are in pyproject.toml — this file
# is kept for simple `pip install -r requirements.txt` workflows.

# --- Core ---
requests>=2.31
yfinance>=0.2.31
pytz>=2024.1

# --- API server ---
fastapi>=0.104
uvicorn[standard]>=0.24
apscheduler>=3.10,<4

# --- Database (PostgreSQL) ---
sqlalchemy[asyncio]>=2.0
asyncpg
alembic

# --- Auth ---
python-jose[cryptography]
passlib[bcrypt]
python-multipart
authlib>=1.3
httpx>=0.25
resend>=2.0

# --- Journal (adapted from Fundamental) ---
pandas>=2.2
numpy>=1.26
ccxt>=4
aiofiles
Pillow

# --- Utilities ---
pydantic-settings
email-validator

# --- Dev (uncomment when needed) ---
# pytest>=7.0
# pytest-asyncio>=0.21
# httpx>=0.25
# ruff>=0.1
